+++ test/queue.php
@@ -1,0 +1,5 @@
+<?php
+
+describe('Sync', function() {
+    require_once __DIR__ . '/queue/sync.php';
+});